+++ oaicli/oai_wrappers.py
@@ -5,7 +5,9 @@ from .oai import (
     get_assistants,
     list_threads,
     create_thread,
-    create_message,
+    list_files,
+    upload_file,
+    save_local_message,
     create_run,
     wait_for_or_cancel_run,
     get_messages,
@@ -14,7 +16,8 @@ from .oai import (
     save_instructions,
     client,
 )
-from . import ASCII_ART, FilePathType
+from . import FilePathType
+import os
 
 
 def create_agent_interactive():
@@ -27,30 +30,114 @@ def create_agent_interactive():
     )
     if input_type == "m":
         instructions = click.prompt("Instructions")
-
     else:
         file_path = click.prompt("Please enter a file path", type=FilePathType())
         click.echo(f"Loading filepath {file_path}")
+        file = os.open(file_path, "r")
+        instructions = file.read()
     new_assistant = create_assistant_wrapper(name=name, instructions=instructions)
     save_instructions(new_assistant, new_assistant.instructions)
     click.echo(f"created {new_assistant.id} ({new_assistant.name})")
     return new_assistant
 
 
-def update_agent_with_file(assistant_id, file_path):
+def select_assistant():
+    assistants = list_assistants()
+    current_assistant = None
+    if len(assistants) > 0:
+        assistant_choice = int(click.prompt("Select Assistant"))
+        current_assistant = assistants[assistant_choice]
+    else:
+        exit("must create assistant")
+    return current_assistant
+
+
+def update_agent():
+    """Update an agents instructions or file list"""
+    assistant = select_assistant()
+    current_assistant = assistant
+    current_assistant_id = assistant.id
+
+    if click.confirm(f"Update prompt?"):
+        filepath = f"{_get_assistant_path(current_assistant)}/instructions.txt"
+        if click.confirm(f"did you update {filepath}?"):
+            new_instructions = None
+        else:
+            new_instructions = click.prompt("enter new instructions")
+        update_agent_instructions(
+            current_assistant_id, new_instructions=new_instructions
+        )
+
+    if click.confirm(f"Update agents files?"):
+        # select file from list of files
+        update_agent_with_file(current_assistant_id)
+
+
+def select_thread():
+    thread = None
+    threads = list_threads()
+
+    if threads and len(threads) == 0:
+        click.echo("No current threads.")
+
+    if click.confirm("Create new thread?"):
+        thread_name = click.prompt("Thread name?")
+        new_thread = create_thread(thread_name=thread_name)
+        threads.append(new_thread)
+
+    # choose thread to send message to
+    for index, thread in enumerate(threads):
+        click.echo(f"{index}. {thread[0]}")
+
+    thread_choice = int(click.prompt("Choose a thread."))
+    current_thread = threads[thread_choice]
+    return current_thread
+
+
+def choose_or_create_file():
+    if click.confirm("Create new file?"):
+        file_path = file_path = click.prompt(
+            "Please enter a file path", type=FilePathType()
+        )
+        if click.confirm(f"Are you sure you want to upload {file_path}?"):
+            file = upload_file(file_path)
+            file_id = file.id
+    else:
+        file_id = select_file_id()
+    return file_id
+
+
+def update_agent_with_file(assistant_id):
     # get agent
     my_assistant = client.beta.assistants.retrieve(assistant_id)
 
-    # list files current agent files
-
-    # get list of local files
-
-    # choose file to update
-
-    # add file id to list
-
+    my_assistant.file_ids += choose_or_create_file()
     # update agent
-    pass
+    client.beta.assistants.update(
+        my_assistant.id,
+        instructions=my_assistant.instructions,
+        name=my_assistant.name,
+        tools=my_assistant.tools,
+        model=my_assistant.modal,
+        file_ids=my_assistant.file_ids,
+    )
+
+
+def run_thread(current_thread_id, current_assistant_id):
+    # run job
+    current_run = create_run(current_thread_id, current_assistant_id)
+    current_run_id = current_run.id
+
+    # wait for result
+    result = wait_for_or_cancel_run(current_thread_id, current_run_id)
+    if not result:
+        return
+
+    # when complete get thread/messages
+    thread_messages = get_messages(current_thread_id)
+    for content in thread_messages[0].content:
+        save_local_message(thread_message=thread_messages[0], role="assistant")
+        click.echo(content.text.value)
 
 
 def update_agent_instructions(assistant_id, new_instructions=None):
@@ -60,8 +147,10 @@ def update_agent_instructions(assistant_id, new_instructions=None):
     click.echo(f"current instructions:\n {my_assistant.instructions}")
 
     if new_instructions:
+        # save them
         save_instructions(my_assistant, new_instructions)
     else:
+        # they are updated locally
         new_instructions = load_instructions(my_assistant)
 
     if click.confirm(f"new instructions OK? \n{new_instructions}"):
@@ -83,10 +172,26 @@ def update_agent_tools(assistant_id, tools_to_add=None, tools_to_remove=None):
 
 
 def select_file_id():
-    pass
+    all_files = list_files()
+    if len(all_files) > 0:
+        choices = [
+            (file.filename, file.id)
+            for file in all_files
+            if file.purpose == "assistants"
+        ]
+        indexes = []
+        for index, (filename, fileid) in choices:
+            indexes.append(0)
+            click.echo(f"{index}. {filename}")
+
+        index_choice = click.prompt("Choose a file index", type=click.Choice(indexes))
+
+        file_choice = choices[index_choice]
+        click.echo(f"You chose file {file_choice[0]}")
+        return file_choice[1]
 
 
-def list_agents():
+def list_assistants():
     assistants = []
     click.echo("Listing assistants...")
     assistant_count = 0